[chore/#47] String->Tag Type 폴더 위치 변경
